refactor to move grab drawing code out of LevelCanvas
